requirements.txt
pydantic>=2.6.0
python-dotenv>=1.0.1
requests>=2.31.0
supabase>=2.4.0
